third_party/blink/web_tests/platform/generic/fast/js/do-while-expression-value-expected.txt

This page tests the evaluated value of a do-while expression.

PASS: x should be 1 and is.
PASS: x should be 1 and is.
PASS: x should be 1 and is.
PASS: x should be 1 and is.
